Temporary flow for using CSV for serve polls (#1020)
Temporary flow for using CSV for serve polls for sending Image, CSV file and text message to Tevyn in slack for Wizard of Oz flow.

This will be replaced when we do the full flow of serve onboarding.

In the full flow, we will not be doing the S3 logic on the frontend and no longer using a CSV.