ENG-4819 add polls onboarding sworn in date and success (#1098)
add polls onboarding sworn in date and success

Related PR: thegoodparty/gp-api#763

Changes:

Updated logic for handling steps in OnboardingPage
This logic was needed because we were using branching logic, but product changed their minds. I included the ability for branching logic because it's more robust as we continue to add more to onboarding.
Added concept of validation to onboarding flow to prevent continuing a step until a certain value is met. This is decently flexible, but if we get to a certain point, we should consider form validation.
Move PreviewCard to high level component directory instead of inside steps which is only for steps of the onboarding flow
Added SuccessPage with auto-redirect after you complete your onboarding
Added DateInputCalendar which uses the newly added Calendar component from goodparty-styleguide
Added Analytics
Hooked up the flow for SwornInDate which sends in API payload and also changes the text message content depending on if date is in present or past.